backend/agrovera_microservices/auth_service/requirements.txt
Django>=5.1
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
